requirements.txt
flask
pandas
scikit-learn
scikit-surprise
joblib
tqdm
flask-cors